Fixed rendering of scales and viewport given in scale will now result in the zoom level assigned to the closest scale available (GEO-154)

+++ util/src/main/java/org/geojsf/util/component/GeoJsfScalesUtil.java
@@ -25,7 +25,7 @@ public class GeoJsfScalesUtil
 	public String getScaleList()
 	{
 		StringBuffer sb = new StringBuffer();
-		for(double d : list)
+		for(Integer d : list)
 		{
 			sb.append(d);
 			sb.append(", ");
